Ya boi, knows what's up with deez linked lists
Added a linked list to manage the input when sending to round robin
method. This can probably be reused for implementing first come first
serve method and shortest job first method if you guys want to. Now that
all the data can be managed correctly, and it works dynamically so that
we don't have to predict the number of processes, all that's lef for
round robin is playing with the counters and print statements.